pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "eventlet>=0.40.4",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "flask-socketio>=5.5.1",
    "google-genai>=1.52.0",
    "gunicorn>=23.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "werkzeug>=3.1.3",
]
